+++ src/test/java/wooteco/subway/station/StationDaoTest.java
@@ -7,9 +7,15 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.dao.DataAccessException;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.line.Line;
+import wooteco.subway.line.LineDao;
+import wooteco.subway.section.SectionDao;
+
+import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest
 @Transactional
@@ -18,6 +24,12 @@ class StationDaoTest {
 
     @Autowired
     private StationDao stationDao;
+
+    @Autowired
+    private LineDao lineDao;
+
+    @Autowired
+    private SectionDao sectionDao;
 
     @DisplayName("역 이름이 주어지면 저장하고 역 객체를 반환한다")
     @Test
@@ -73,5 +85,29 @@ class StationDaoTest {
 
         Station foundStation = stationDao.findById(station.getId());
         assertThat(foundStation.getName()).isEqualTo(stationName);
+    }
+
+    @DisplayName("노선 id를 통해 노선에 포함된 역의 id들을 조회한다")
+    @Test
+    void findStationsIdByLineId() {
+        String stationName1 = "강남역";
+        String stationName2 = "잠실역";
+        String stationName3 = "신림역";
+
+        Station station1 = stationDao.save(stationName1);
+        Station station2 = stationDao.save(stationName2);
+        Station station3 = stationDao.save(stationName3);
+
+        final Long stationId1 = station1.getId();
+        final Long stationId2 = station2.getId();
+        final Long stationId3 = station3.getId();
+
+        String name = "2호선";
+        String color = "green";
+        Line savedLine = lineDao.save(name, color);
+        sectionDao.save(savedLine.getId(), stationId1, stationId2);
+        sectionDao.save(savedLine.getId(), stationId2, stationId3);
+
+        assertTrue(stationDao.findStationsIdInLineId(savedLine.getId()).containsAll(Arrays.asList(stationId1, stationId2, stationId3)));
     }
 }
